WEKA CLONE
The aim of this project is to provide users with a user-friendly tool for uploading datasets, performing data preprocessing tasks such as handling missing values, and applying various machine learning algorithms for classification tasks. It aims to simplify the process of data analysis and model building by offering a graphical interface and easy-to-use functionalities.